Async fixture download
